Update when-to-deploy-on-k8s.md
Removed abbreviations - unless we clearly state what we mean by SQS and K8, we should not introduce these abbreviations.

Removed "however" - for clarity it did not seem that the first sentence after "-IMPORTANT" was a continuation of line 3

Removed parenthesis - for ease of reading.  First pass through this required me to re-read that section to make it make sense.  

Modified the second to last and 3rd to last sentences to be more direct "we support SQ only, we are not responsible for your customizations"